+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+Features:
+- Add checked icon to `Button` component.
+
 ## [2.7.0] - 2019-05-24
 
 Features: